+++ src/Events/Tag/Tag.php
@@ -11,6 +11,9 @@ use function microtime;
 /** @internal */
 abstract class Tag implements Command
 {
+    public const TAG_STACK_TRACE = 'stack';
+    public const TAG_ARGUMENTS   = 'desc';
+
     /** @var RequestId */
     protected $requestId;